+++ src/main/java/com/barlinc/unusual_prehistory/blocks/WaterEggBlock.java
@@ -2,6 +2,7 @@ package com.barlinc.unusual_prehistory.blocks;
 
 import com.barlinc.unusual_prehistory.blocks.blockentity.ExtraDataBlockEntity;
 import com.barlinc.unusual_prehistory.blocks.blockentity.WaterEggBlockEntity;
+import com.barlinc.unusual_prehistory.entity.KimmeridgebrachypteraeschnidiumNymph;
 import com.barlinc.unusual_prehistory.entity.base.PrehistoricAquaticMob;
 import com.barlinc.unusual_prehistory.entity.base.PrehistoricMob;
 import net.minecraft.advancements.CriteriaTriggers;
@@ -55,6 +56,9 @@ public class WaterEggBlock extends WaterEggBlockEntity {
                     }
                     if (entity instanceof PrehistoricMob prehistoricMob) {
                         prehistoricMob.setFromEgg(true);
+                    }
+                    if (entity instanceof KimmeridgebrachypteraeschnidiumNymph nymph) {
+                        nymph.setFromEgg(true);
                     }
                     int k = random.nextInt(1, 361);
                     if (placer != null) {